+++ builds/compromise.js
@@ -1,4 +1,4 @@
-/* compromise 13.10.3 MIT */
+/* compromise 13.10.4 MIT */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -2597,9 +2597,9 @@
   // prefixes: ! [ ^
   //     [\!\[\^]*
   // match  'foo /yes/' and not 'foo/no/bar'
-  var bySlashes = /(?:^|\s)([\!\[\^]*(?:<[^<]*>)?\/.*?[^\\\/]\/[\?\]\+\*\$~]*)(?:\s|$)/g; // match '(yes) but not foo(no)bar'
+  var bySlashes = /(?:^|\s)([\!\[\^]*(?:<[^<]*>)?\/.*?[^\\\/]\/[\?\]\+\*\$~]*)(?:\s|$)/; // match '(yes) but not foo(no)bar'
 
-  var byParentheses = /(?:^|\s)([\!\[\^]*(?:<[^<]*>)?\(.*?[^\\\)]\)[\?\]\+\*\$~]*)(?:\s|$)/g; // okay
+  var byParentheses = /([\!\[\^]*(?:<[^<]*>)?\([^\)]+[^\\\)]\)[\?\]\+\*\$~]*)(?:\s|$)/; // okay
 
   var byWord = / /g;
 
@@ -2627,6 +2627,11 @@
     var res = []; // parse by (blocks), next
 
     arr.forEach(function (str) {
+      if (isReg(str)) {
+        res.push(str);
+        return;
+      }
+
       res = res.concat(str.split(byParentheses));
     });
     res = cleanUp(res); // split by spaces, now
@@ -2645,7 +2650,7 @@
     return _final;
   };
 
-  var _01ParseBlocks = parseBlocks; // console.log(parseBlocks(`[<num>#Value] [<currency>(mark|rand|won|rub|ore)] foo`))
+  var _01ParseBlocks = parseBlocks; // console.log('(one two) (upto) [<snooze_to>#Date+]'.split(byParentheses))
 
   /* break-down a match expression into this:
   {
@@ -3861,7 +3866,7 @@
 
   var fromJSON_1 = fromJSON;
 
-  var _version = '13.10.3';
+  var _version = '13.10.4';
 
   var entity = ['Person', 'Place', 'Organization'];
   var nouns$1 = {